feat(account): extract AccountSheet + PLAY hub/arena polish
Founder feedback round 2 on the Play Kingdom hub:

1. Extract AccountSheet from exercises-screen into a reusable
   components/account/account-sheet.tsx (same props/behavior, moved verbatim
   incl. networkName helper). exercises-screen imports it; 5 now-orphaned
   imports removed. It was a local function tied to that screen — now any
   surface can mount the SAME sheet.
2. Wire AccountSheet inline in /arena. The arena account button opened
   /exercises?sheet=account, but play mode cross-mode-redirects /exercises to
   the LEARN deployment (mode-routing.ts), so the sheet never opened in-mode.
   Now arena mounts <AccountSheet> and the button opens it directly (proSheet/
   shopSheet/coach-credits/disconnect wired from arena's existing hooks).
3. Hide the account circle in the PLAY hub (founder: account lives on /arena,
   not the play home). Peones chip stays. Removed onAccountTap prop + wiring.
4. Hide the WarmUp modal on /arena. Once it collapsed to a single "ARENA" CTA
   it added no decision value; SoftGateSheet + wiring stay intact (re-enable by
   restoring setSoftGateOpen(!hasAnyPieceProgress())).

Issue "PRO not updating on MiniPay" was NOT a code bug — the ngrok origin
wasn't allowlisted; resolved by setting NEXT_PUBLIC_APP_URL/PREVIEW_URL to the
tunnel. enforceOrigin untouched.

Typecheck clean; 232 hub/arena/exercises tests green; build EXIT=0. Drove /hub
(no account circle) + /arena?fresh=1 (no WarmUp modal) at 390px. Arena account
connected-state not driven (needs wallet); reuses the extracted sheet + tests.

Wolfcito 🐾 @akawolfcito
