+++ apps/web/src/components/hub/play-hub-scaffold.tsx
@@ -21,9 +21,6 @@ type PlayHubScaffoldProps = {
   onCoachTap: () => void;
   onShopTap: () => void;
   onArenaPress: () => void;
-  /** Opens the account surface (same universal sheet as LEARN + arena —
-   *  `/exercises?sheet=account`). Only wired/rendered when connected. */
-  onAccountTap: () => void;
 };
 
 /** Play Kingdom home. Mirrors the LEARN/LITE hub's visual system (unified
@@ -40,12 +37,10 @@ export function PlayHubScaffold({
   onCoachTap,
   onShopTap,
   onArenaPress,
-  onAccountTap,
 }: PlayHubScaffoldProps) {
   const tHud = useTranslations("HUD_COPY");
   const tRail = useTranslations("HUB_ACTION_RAIL_COPY");
   const tPlay = useTranslations("PLAY_HUB_COPY");
-  const tStatus = useTranslations("GLOBAL_STATUS_BAR_COPY");
   const proAriaLabel = pro.active
     ? tHud("proAriaLabel", { days: pro.daysRemaining })
     : tHud("proInactiveAriaLabel");
@@ -74,36 +69,9 @@ export function PlayHubScaffold({
             <LanguageChip />
           </div>
           <div className="hub-scaffold-hud-right">
-            {/* Account entry — same circular avatar chip + universal account
-                sheet as the LEARN header (Daily gift is Lite-only, so PLAY's
-                top-right anchor stays the PRO badge below instead). */}
-            {isWalletConnected ? (
-              <button
-                type="button"
-                onClick={onAccountTap}
-                aria-label={
-                  pro.active
-                    ? tStatus("proManageLabel")
-                    : tStatus("accountLabel")
-                }
-                data-testid="play-hub-account-chip"
-                className={`hub-account-circle${
-                  pro.active ? " hub-account-circle--pro" : ""
-                }`}
-              >
-                {/* eslint-disable-next-line jsx-a11y/aria-unsupported-elements */}
-                <picture className="hub-account-circle-avatar">
-                  <source srcSet="/art/avatar-small-account.avif" type="image/avif" />
-                  <source srcSet="/art/avatar-small-account.webp" type="image/webp" />
-                  <img
-                    src="/art/avatar-small-account.png"
-                    alt=""
-                    aria-hidden="true"
-                    draggable={false}
-                  />
-                </picture>
-              </button>
-            ) : null}
+            {/* Account entry intentionally omitted here (founder 2026-07-07):
+                the PLAY hub keeps only trophy·Peones·language + PRO badge. The
+                account surface is reachable from /arena instead. */}
             {!isWalletConnected ? (
               <button
                 type="button"
